# InMemoryLog fixes