Some fixes to the InMemoryLog, specifically concurrency and resizing mechanism. This should address #152 and #153.
Although this is a low priority, this is something I was playing with. And it might come in hand for the ISPN counters: 

https://github.com/infinispan/infinispan/blob/c2db1cc26bcebe2a4edc3ac583bda07948aa1fb8/core/src/main/java/org/infinispan/remoting/transport/jgroups/JGroupsRaftManager.java#L115-L119

All in all, it does pretty much the same thing as before but instead uses the ArrayRingBuffer and a read-write lock instead of synchronized blocks. For the synchronization part, I am a bit unsure if it is actually needed, as we consume messages in a single thread maybe we can get rid of the lock. But since I was not completely sure, I left it as is.

Running this version against main for the counter-perf we have:

```text
NEW:
======================= Results: ===========================
A: 168,343.58 updates/sec (20,203,587 updates, 593.72us / update)
B: 128,772.41 updates/sec (15,454,363 updates, 775.97us / update)


Throughput: 148,558.08 updates/sec/node
Time:       672.71us / update

MAIN:
======================= Results: ===========================
A: 10,831.18 updates/sec (1,300,879 updates, 9.22ms / update)
B: 9,998.01 updates/sec (1,201,111 updates, 10.00ms / update)


Throughput: 10,414.54 updates/sec/node
Time:       9.60ms / update
```

TBH these numbers make me think I did something wrong :sweat_smile:
But I consistently got values like these in multiple executions. I am running for 2 min, sync version, with 2 nodes. The leader has 3 cores and the follower has another 3 cores, so they shouldn't compete. I didn't run the LogJmhBenchmark for comparison between versions, as it is specific for appending entries I would expect both versions to have similar performance (another reason why the counter-perf makes me worried).

I ran it for 30min without OOM, and for 1 min capturing with JMC. Both leader and follower have a similar graph:

![image](https://user-images.githubusercontent.com/13581903/211958907-a0afff8a-8c6f-4a62-a7c7-7f866e9352c9.png)

This is definitely a low-priority PR but the counter-perf numbers made me think it could be worth opening the PR (as the changes aren't that big) or at least starting a discussion.