@belaban 2023-01-12
How did you run the tests? CounterPerf with `InMemoryLog`? The old impl could run out of memory during long streches without truncation.
I'm not going to look into details, I trust you did the right thing...