@jabolina 2023-01-12
> How did you run the tests? CounterPerf with `InMemoryLog`? 

Yes, I changed the configuration to use the `InMemoryLog`, and ran the java command directly without using `./bin/counter-perf`, so I also increased Xmx. For the 2 min runs, the main version handled without throwing exceptions.

We can close #152 but looks like I don't have permission to do it, tough.