@belaban 2023-01-13
I closed #152, thanks. I also added you to collabs, so you should be able to close issues yourself.