@pruivo 2023-01-13
@belaban @jabolina is the read-write lock still required? Isn't RAFT single-threaded now?