@belaban 2023-01-14
Shouldn't an uncontended lock have more or less no cost?