@pruivo 2023-01-16
> Shouldn't an uncontended lock have more or less no cost?

AFAIK yes, but zero cost is better than low cost :)
Also, I'm concerned about the lambda (`Function`) allocations in every operation:

```java
    public <V> V read(Function<T, V> function) {
        Lock rl = lock.readLock();
        try {
            rl.lock();
            return function.apply(value);
        } finally {
            rl.unlock();
        }
    }

    public <V> V write(Function<T, V> function) {
        Lock wl = lock.writeLock();
        try {
            wl.lock();
            return function.apply(value);
        } finally {
            wl.unlock();
        }
    }
```